Register user status to auth store
