+++ frontend/src/scripts/ui/layout/Header.ts
@@ -1,38 +1,7 @@
 import { logout } from "../../modules/auth/auth.service";
 import { authStore } from "../../modules/auth/auth.store";
 import { createEl } from "../../utils/dom-helper";
-
-const showUserStatus = async (container: HTMLElement): Promise<Boolean> => {
-    const loggedUser = authStore.get().username;
-
-    if (!loggedUser) {
-        window.log.error("No user logged in.");
-        return false;
-    }
-
-    // Create status wrapper
-    const statusWrapper = document.createElement("div");
-    statusWrapper.className =
-        "absolute items-center justify-between bg-gray-100 rounded p-2 shadow-sm right-4 top-4";
-
-    // User display name
-    const userNameEl = document.createElement("span");
-    userNameEl.textContent = `Logged in as ${loggedUser}`;
-    userNameEl.className = "text-gray-700 text-sm";
-
-    // Logout button
-    const logoutBtn = document.createElement("button");
-    logoutBtn.textContent = "Logout";
-    logoutBtn.className =
-        "ml-4 bg-red-500 hover:bg-red-600 text-white text-sm py-1 px-3 rounded focus:outline-none";
-
-    logoutBtn.onclick = logout;
-
-    statusWrapper.appendChild(userNameEl);
-    statusWrapper.appendChild(logoutBtn);
-    container.appendChild(statusWrapper);
-    return true;
-};
+import { createUserStatus } from "./UserStatus";
 
 export const createHeader = async (): Promise<HTMLElement> => {
     const anchor = createEl("a", "text-3xl font-bold", {
@@ -62,7 +31,18 @@ export const createHeader = async (): Promise<HTMLElement> => {
 
     const loginStatus = createEl("li");
     navList.appendChild(loginStatus);
-    await showUserStatus(loginStatus); // FIX
+
+    const unsubscribe = authStore.subscribe((state) => {
+        window.log.debug("AuthStore triggered in header");
+        if (state.isAuthenticated && state.username) {
+            createUserStatus(loginStatus, state.username);
+        }
+    });
+
+    document.addEventListener("destory", () => {
+        window.log.debug("Header unsubscribe to Login status");
+        unsubscribe();
+    });
 
     const nav = createEl("nav", "flex items-center space-x-6", { children: [navList] });
